delete popUps

+++ client/src/app/features/categories/category-details/category-recipes.component.ts
@@ -3,6 +3,7 @@ import { RecipesCategory } from 'app/core/models/RecipeCategory';
 import { Pagination } from '../../../core/helpers/Pagination';
 import { SharedService } from '../../../core/shared.service';
 import { ToastrService } from 'ngx-toastr';
+import { NgxBootstrapConfirmService } from 'ngx-bootstrap-confirm';
 
 @Component({
   selector: 'app-category-recipes',
@@ -10,41 +11,49 @@ import { ToastrService } from 'ngx-toastr';
   styleUrls: ['./category-recipes.component.css']
 })
 export class CategoryRecipesComponent implements OnInit {
-  
+
   CategoryRecipesList: RecipesCategory[];
-  pagination: Pagination; 
-  pageNumber=1; 
-  pageSize=10; 
-  constructor(private service:SharedService, public toastr:ToastrService) { }
+  pagination: Pagination;
+  pageNumber = 1;
+  pageSize = 10;
+  constructor(private service: SharedService, public toastr: ToastrService, private ngxBootstrapConfirmService: NgxBootstrapConfirmService) { }
 
   ngOnInit(): void {
     this.loadCategories();
   }
 
 
-  loadCategories(){
-    this.service.getCategories(this.pageNumber, this.pageSize).subscribe(response=>{
-      this.CategoryRecipesList= response.result; 
-      this.pagination= response.pagination; 
+  loadCategories() {
+    this.service.getCategories(this.pageNumber, this.pageSize).subscribe(response => {
+      this.CategoryRecipesList = response.result;
+      this.pagination = response.pagination;
     });
   }
 
-  pageChanged(event:any)
-  {
-    this.pageNumber= event.page;
-    this.loadCategories(); 
+  pageChanged(event: any) {
+    this.pageNumber = event.page;
+    this.loadCategories();
   }
 
-  deleteCategory(item){
-    this.service.deleteCategory(item).subscribe(data=>{
-      this.loadCategories();
+  deleteCategory(item) {
+    let options = {
+      title: 'Sure you want to delete this category?',
+      confirmLabel: 'Yes',
+      declineLabel: 'Cancel'
+    }
+    this.ngxBootstrapConfirmService.confirm(options).then((res: boolean) => {
+      if (res) {
+        this.service.deleteCategory(item).subscribe(data => {
+          this.loadCategories();
+        })
+        setTimeout(() => {
+          location.reload();
+        }, 1000);
+      }
+      this.toastr.success("Category succesfully deleted");
     }), err => {
       this.toastr.error(err.error);
       console.log("Unable to delete category");
     }
-    this.toastr.success("Category succesfully deleted");
-    setTimeout(() => {
-      location.reload();
-    }, 2000);
   }
 }
